+++ app/Http/Livewire/Task/SingleTask.php
@@ -102,7 +102,9 @@ class SingleTask extends Component
         foreach ($this->task->images ?? [] as $image) {
             Storage::delete($image);
         }
-        $this->task->oembed->delete();
+        if ($this->task->oembed) {
+            $this->task->oembed->delete();
+        }
         $this->task->delete();
         $this->emitUp('refreshTasks');
