+++ app/Http/Livewire/Tasks/SingleTask.php
@@ -63,7 +63,9 @@ class SingleTask extends Component
         foreach ($this->task->images ?? [] as $image) {
             Storage::delete($image);
         }
-        $this->task->oembed->delete();
+        if ($this->task->oembed) {
+            $this->task->oembed->delete();
+        }
         $this->task->delete();
 
         return $this->emitUp('refreshTasks');
